refactor: combine cmdline_controller with cmdline_manager
